Solve duplicate file match issue.
OS X HFS+ is case insensitive by default. When we glob for a file that
doesn't match its case-sensitive representation, we get multiple
results, and the set doesn't eliminate duplicates.

Rather than use a case-insensitive set or upper/lower all of our
matches, just test for a file using os.listdir which returns the
case-sensitive version of a filename.