Change naming conventions to more accurately reflect what is targeted.
This project began as a way to detect and remove adware defined by Apple
in the KnowledgeBase article HT203987 located at
https://support.apple.com/en-us/ht203987. As such, the original code and
accompanying documentation used the term Adware to refer to the software
and files targeted. However, as usage of SavingThrow grew, legitimate
software that admin's desired to detect and remove from their enterprise
machines were included in the potential ADF files used as data to drive
SavingThrow. The labelling of such software as Adware was inaccurate and
has been corrected in this commit to describe that software simply as
"Apps". It is up to the individual discretion of the administrator's
making use of this software to determine whether a piece of software is
indeed desired or not. No reflection the software targeted is intended
or meant by the historical use of the term "Adware" in this software,
and any continued presence of that term, or anything else negative is
either unintentional, overlooked (please submit an issue for us to
resolve it!), or left in place to support existing users and the ADF
file format, which is used more broadly than SavingThrow, and is no
longer maintained or affiliated with the author of this project.